test: create test utils functions
