+++ src/__tests__/Migration.test.ts
@@ -1,28 +1,11 @@
 import { Logger } from '@adonisjs/logger/build/standalone';
 import { ClientSession } from 'mongodb';
 
+import { getMongodb, getLogger } from '../../test-utils/TestUtils';
 import createMigration from '../Migration';
-import { Mongodb } from '../Mongodb';
 
-const mongoConfig = {
-  default: 'mongo',
-  connections: {
-    mongo: {
-      url: 'mongodb://127.0.0.1:33333',
-      database: 'test-runner',
-    },
-  },
-};
-
-const loggerConfig = {
-  name: 'adonis-logger',
-  level: 'trace',
-  messageKey: 'msg',
-  enabled: true,
-};
-
-const logger = new Logger(loggerConfig);
-const db = new Mongodb(mongoConfig, logger);
+const logger = getLogger();
+const db = getMongodb(logger);
 const BaseMigration = createMigration(db);
 
 class TestMigration extends BaseMigration {
@@ -39,7 +22,7 @@ class TestMigration extends BaseMigration {
 }
 
 afterAll(async () => {
-  await (await db.connection('mongo').database()).dropDatabase();
+  //await (await db.connection('mongo').database()).dropDatabase();
   await db.closeConnections();
 });
