+++ src/__tests__/Mongodb.test.ts
@@ -1,25 +1,8 @@
 import { Logger } from '@adonisjs/logger/build/standalone';
 
-import { Mongodb } from '../Mongodb';
+import { getMongodb } from '../../test-utils/TestUtils';
 
-const mongoConfig = {
-  default: 'mongo',
-  connections: {
-    mongo: {
-      url: 'mongodb://127.0.0.1:33333',
-      database: 'test-runner',
-    },
-  },
-};
-
-const loggerConfig = {
-  name: 'adonis-logger',
-  level: 'trace',
-  messageKey: 'msg',
-  enabled: false,
-};
-
-const db = new Mongodb(mongoConfig, new Logger(loggerConfig));
+const db = getMongodb();
 
 afterAll(async () => {
   await db.connection('mongo').close();
